feat: Introduce MetaClusterModel with candidate cluster generation and hierarchical structure for meta-clustering
